Refactor AddWorker method to remove DbContext dependency
The generic type parameter `<DbContext>` was removed from the
`AddWorker` method in `ServiceCollectionExtensions` to simplify
its usage and make it more general-purpose. The method signature
was updated from `AddWorker<DbContext>(this IServiceCollection services)`
to `AddWorker(this IServiceCollection services)`.

This change decouples the method from a specific `DbContext` type,
allowing it to be used in scenarios where a `DbContext` is not
required. The addition of the `EventBusWorker` as a hosted service
remains unchanged.